Implement all CRUD functions
/add, /view, /edit and /delete commands are implemented.

However, ingredients and steps cannot be named "<< back" due to the
callback_data of the "<< back" button. Steps are also
allowed to be identical to each other, meaning that deleting one copy
will delete all other copies. Moreover, steps can only be added to the
end of the recipe for now.