Allow viewing of recipes from other users
Recipes can be toggled between private and public statuses.
Public recipes can be viewed by anyone as long as they have
the username of the user. Recipes of users with no username
cannot be viewed by others as well.

Any changes to the username of a user will be updated accordingly
in the database at 12pm daily.